Add player class and race selection feature
Introduced the `Player` class to model player characters using the `dataclass` decorator. Implemented a new method in the `Game` class to build a dictionary of races from the game database and allow players to select a race at the start of the game. Enhanced type hinting in database parsing functions and added a safeguard to check database integrity before starting the game.